pet-object-detection/requirements.txt
torch
numpy
pydantic
daily-python
ultralytics
opencv-python
pillow